Separation of all types into separate files & Fixes
