+++ src/app/api/auth/signin/route.ts
@@ -1,9 +1,9 @@
-import { IEnterUserData, TUser } from "@/interfaces/interfaces";
 import db from "@/lib/db";
 import jwt from "jsonwebtoken";
 import { NextResponse } from "next/server";
 import { compare } from "bcrypt";
 import { cookies } from "next/headers";
+import { IEnterUserData, TUser } from "@/interfaces/userType";
 
 export const POST = async (req: Request, res: Response) => {
   const data: IEnterUserData = await req.json();
