+++ src/interfaces/poolSettingType.ts
@@ -1,0 +1,7 @@
+export interface IPoolSettings {
+  host?: string;
+  port?: number;
+  database?: string;
+  user?: string;
+  password?: string;
+}
