+++ src/interfaces/responseType.ts
@@ -1,0 +1,5 @@
+export interface IResponse {
+  message: string;
+  status: number;
+  data: IUser | null;
+}
