+++ src/interfaces/sessionType.ts
@@ -1,0 +1,4 @@
+export interface ISession {
+  id: number;
+  email: string;
+}
